48 change responseid type (#49)
* Refactor ResponseId to use MessageId enumeration

Updated ResponseId properties across multiple files to use the MessageId type instead of byte. This change enhances type safety and clarity in message identification. Additionally, some ResponseId assignments were modified to cast to uint where applicable.

* Update AntUsbStick to version 3.1.1.0

- Added new version (3.1.1.0) to version history.
- Updated release notes to reflect changes in AntResponse ResponseId property type.
- Modified ContentLayout.content to include the new version as a selectable topic.
- Updated AntUsbStick.csproj with the new assembly version.
- Created v3.1.1.0.aml with structured documentation, including an introduction and related topics.

* Update to version 4.0.0.0 with breaking changes

- Changed assembly version from 3.0.0.0 to 4.0.0.0 in `AntRadioInterface.csproj`.
- Updated package release notes to reflect breaking change in `ResponseId` property type from `byte` to `MessageId`.
- Added new version entry for 4.0.0.0 in `AntRadioInterfaceVersionHistory.aml`.
- Included new topic for version 4.0.0.0 in `ContentLayout.content`.
- Updated `Documentation.shfbproj` to include new version file.
- Created `v4.0.0.0.aml` detailing breaking changes and related topics.

+++ Examples/MAUI-gRPC/AntGrpcService/Services/AntChannelService.cs
@@ -48,7 +48,7 @@ namespace AntGrpcService.Services
                     ChannelNumber = channelResponse.ChannelNumber,
                     ThresholdConfigurationValue = channelResponse.ThresholdConfigurationValue,
                     Payload = ByteString.CopyFrom(channelResponse.Payload),
-                    ResponseId = channelResponse.ResponseId,
+                    ResponseId = (uint)channelResponse.ResponseId,
                     Rssi = channelResponse.Rssi,
                     Timestamp = channelResponse.Timestamp
                 });
